+++ typescript/registry/RELEASES.md
@@ -7,3 +7,13 @@ Based on:
 - Speakeasy CLI 1.543.4 (2.598.22) https://github.com/speakeasy-api/speakeasy
 ### Generated
 - [typescript v0.3.0] typescript/registry
+
+## 2025-05-11 05:19:51
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.543.4 (2.598.22) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.3.4] typescript/registry
+### Releases
+- [NPM v0.3.4] https://www.npmjs.com/package/@smithery/registry/v/0.3.4 - typescript/registry